+++ blosc/trunc-prec.h
@@ -14,7 +14,7 @@
 #include <stdio.h>
 #include <stdint.h>
 
-void truncate_precision(uint8_t prec_bits, int32_t typesize, int32_t nbytes,
-                        const uint8_t* src, uint8_t* dest);
+int truncate_precision(int8_t prec_bits, int32_t typesize, int32_t nbytes,
+                       const uint8_t* src, uint8_t* dest);
 
 #endif //BLOSC_TRUNC_PREC_H